refactor: (#100) 테스트 관련 설정
