+++ src/test/java/org/dinosaur/foodbowl/domain/healthcheck/presentation/HealthCheckControllerTest.java
@@ -9,13 +9,12 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import java.util.Optional;
-import org.dinosaur.foodbowl.PresentationTest;
 import org.dinosaur.foodbowl.domain.auth.jwt.JwtTokenProvider;
 import org.dinosaur.foodbowl.domain.healthcheck.application.HealthCheckService;
 import org.dinosaur.foodbowl.domain.healthcheck.dto.response.HealthCheckResponse;
 import org.dinosaur.foodbowl.domain.member.domain.Member;
 import org.dinosaur.foodbowl.domain.member.domain.vo.SocialType;
-import org.dinosaur.foodbowl.domain.member.persistence.MemberRepository;
+import org.dinosaur.foodbowl.test.PresentationTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
@@ -31,9 +30,6 @@ class HealthCheckControllerTest extends PresentationTest {
 
     @Autowired
     private JwtTokenProvider jwtTokenProvider;
-
-    @MockBean
-    private MemberRepository memberRepository;
 
     @MockBean
     private HealthCheckService healthCheckService;
