+++ src/test/java/org/dinosaur/foodbowl/test/IntegrationTest.java
@@ -1,0 +1,12 @@
+package org.dinosaur.foodbowl.test;
+
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.transaction.annotation.Transactional;
+
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
+@Transactional
+@SpringBootTest
+public class IntegrationTest {
+}
